+++ ops/scripts/ansible_init.sh
@@ -66,6 +66,12 @@ echo "fuw_db_user = $fuw_db_user" >> ansible.properties
 echo "fuw_db_password = $fuw_db_password" >> ansible.properties
 echo "fuw_db_database = $fuw_db_database" >> ansible.properties
 
+# Required by rclone.conf.j2 for bastion server to copy readme files to Wasabi bucket
+wasabi_access_key_id=$(curl -s --header "PRIVATE-TOKEN: $GITLAB_PRIVATE_TOKEN" "$PROJECT_VARIABLES_URL/WASABI_ACCESS_KEY_ID?filter%5benvironment_scope%5d=$target_environment" | jq -r .value)
+wasabi_secret_access_key=$(curl -s --header "PRIVATE-TOKEN: $GITLAB_PRIVATE_TOKEN" "$PROJECT_VARIABLES_URL/WASABI_SECRET_ACCESS_KEY?filter%5benvironment_scope%5d=$target_environment" | jq -r .value)
+echo "wasabi_access_key_id = $wasabi_access_key_id" >> ansible.properties
+echo "wasabi_secret_access_key = $wasabi_secret_access_key" >> ansible.properties
+
 echo "deployment_target = $deployment_target" >> ansible.properties
 echo "gitlab_project = $gitlab_project" >> ansible.properties
 echo "ssh_private_key_file = $aws_ssh_key" >> ansible.properties
@@ -82,7 +88,7 @@ echo "aws_secret_access_key = $aws_secret_access_key" >> ansible.properties
 bastion_ip=$(terraform output ec2_bastion_public_ip | sed 's/"//g')
 webapp_ip=$(terraform output ec2_private_ip | sed 's/"//g')
 
-echo "ec2_bastion_login_account= centos@$bastion_ip" >> ansible.properties
+echo "ec2_bastion_login_account = centos@$bastion_ip" >> ansible.properties
 
 # variables needed by disk-usage-monitor
 gitter_room_id=$(curl -s --header "PRIVATE-TOKEN: $GITLAB_PRIVATE_TOKEN" "$FORK_VARIABLES_URL/GITTER_IT_NOTIFICATION_ROOM_ID" | jq -r .value)